#1376: tests: split off lb args enum converter tests
